Data from Experiment 1 in Castillo et al. (2024)

Participants produced a random sequence of heights of either men or women in the United Kingdom. In one sequence, they sampled heights as distributed according to a uniform distribution (Uniform condition); in the other sequence, heights were distributed following their actual distribution (which is roughly Gaussian). These data are licensed under CC BY 4.0, reproduced from materials in OSF.

id

participant id

part_Gender

participant's gender (self-reported)

part_Height

participant's own height (self-reported)

part_Home

participant's home country (self-reported)

RQ_Rep

percentage of correct responses in Randomness Questionnaire, for coin toss pairs where one sequence had too many repetitions

RQ_Alt

percentage of correct responses in Randomness Questionnaire, for coin toss pairs where one sequence had too many alternations

RQ_GFM

percentage of correct responses in Randomness Questionnaire, Gambling Fallacies Measure section

minHeight

height participant reports to be the shortest adult in the UK (from target gender)

maxHeight

height participant reports to be the tallest adult in the UK (from target gender)

condition

whether the participant did the uniform condition first (UN) or not (NU)

target_gender

gender they had to generate heights from, either male (M) or female (F)

index

position of the item in the sequence, 0 indexed

block

whether the item belongs to the first sequence the participant uttered (A) or the second (B)

target_dist

whether the instructions asked for heights as distributed in the population (N) or uniformly distributed (U)

label

what the participant uttered

unit

height unit, either centimetres (cm) or feet and inches (f_in).

value

value in cms of the height uttered.

value_in_units

value of the height uttered depending on the value of unit (either in inches or in centimetres). Used to calculate adjacencies, distances, etc.

starts

timestamp of when the utterance starts, in seconds.

delays

temporal difference with the start of the previous item (i.e. starts[index] - starts[index - 1])

R

whether the item is a repetition of the last

A

whether the item is adjacent to the last (after removing repetitions)

TP_full

whether the item is a turning point, considering all items (after removing repetitions)

D

the Euclidean distance to the previous item (after removing repetitions)

S

a measure of how likely the item is in a uniform or gaussian distribution (see text)

expected_*

the expectation for measure * derived from reshuffling the participant's sequence 10000 times

castillo2024.rgmomentum.e1

Format

An object of class data.frame with 5836 rows and 29 columns.

Source

https://osf.io/dw8ez/

References

Castillo L, León-Villagrá P, Chater N, Sanborn AN (2024). “Explaining the Flaws in Human Random Generation as Local Sampling with Momentum.” PLOS Computational Biology, 20(1), 1--24. doi:10.1371/journal.pcbi.1011739 .