Data from Experiment 2 in Castillo et al. (2024)

Participants first learned a set of syllables arranged in either a single row (one-dimensional condition) or a grid (two-dimensional condition), then produced two random sequences for the same display. These data are licensed under CC BY 4.0, reproduced from materials in OSF.

id

participant id

part_Gender

participant's gender (self-reported)

part_Age

participant's age (self-reported)

index

position of the item in the sequence, 0 indexed

id

unique identifier for the participant

block

whether the item belongs to the first sequence the participant uttered (A) or the second (B)

syll

syllable uttered

starts

timestamp of when the utterance starts, in seconds.

delays

temporal difference with the start of the previous item (i.e. starts[index] - starts[index - 1])

dim

whether the participant was allocated to the one-dimensional or two-dimensional condition

seed

Which of five possible configurations the participant learned

position

The position of the syllable in the array. For 1D arrays, position is left to right. For 2D arrays positions 1-2 correspond to the top 2 cells; 3-5 to the middle 3 cells; and 6-7 to the bottom three cells (always left to right)

R

whether the item is a repetition of the last

A

whether the item is adjacent to the last in the display (after removing repetitions)

TP_full

whether the item is a turning point, considering all items (after removing repetitions)

D

the Euclidean distance to the previous item (after removing repetitions)

S

a measure of how likely the item is in a uniform or gaussian distribution (see text)

expected_*

the expectation for measure * derived from reshuffling the participant's sequence 10000 times

castillo2024.rgmomentum.e2

Format

An object of class data.frame with 28483 rows and 20 columns.

Source

https://osf.io/dw8ez/

References

Castillo L, León-Villagrá P, Chater N, Sanborn AN (2024). “Explaining the Flaws in Human Random Generation as Local Sampling with Momentum.” PLOS Computational Biology, 20(1), 1--24. doi:10.1371/journal.pcbi.1011739 .